cleanup: break down serializers into files
