+++ src/persist/serializer/union.ts
@@ -1,39 +1,5 @@
-import { DecodeError } from "./error";
-import { asNumber, asString, minWidth } from "./helper";
-
-export interface Serializer<T> {
-  accepts(value: unknown): value is T;
-  encode(value: T): string;
-  decode(value: string): T;
-}
-
-export const StringSerializer: Serializer<string> = {
-  accepts(value: unknown) {
-    return typeof value === "string";
-  },
-
-  encode(value: string): string {
-    return value;
-  },
-
-  decode(value: string): string {
-    return value;
-  },
-};
-
-export const NumberSerializer: Serializer<number> = {
-  accepts(value: unknown) {
-    return typeof value === "number";
-  },
-
-  encode(value: number): string {
-    return asString(value, minWidth(value));
-  },
-
-  decode(value: string): number {
-    return asNumber(value);
-  },
-};
+import { DecodeError } from "../error";
+import { Serializer } from "./index";
 
 class UnionSerializer<T> implements Serializer<T> {
   private serializers: readonly Serializer<unknown>[];
